requirements.txt
alembic==1.16.5
Authlib==1.6.4
blinker==1.9.0
cffi==2.0.0
click==8.2.1
cryptography==46.0.1
fido2==2.0.0
Flask==3.1.2
flask-cors==6.0.1
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysqlclient==2.2.7
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
